postinstall stuff

+++ scripts/webidl/convert.js
@@ -31,12 +31,19 @@ addDir("../../lib/jsdom/living/domparsing");
 
 const outputDir = path.resolve(__dirname, "../../lib/jsdom/living/generated/");
 
-// Clean up any old stuff lying around.
-rimraf.sync(outputDir);
-fs.mkdirSync(outputDir);
+(function() {
+  // we don't wanna build this on post install since it was already built with
+  // prepublish, but installing from github won't have the generated scripts 
+  // and cding into node_modules to run scripts is messy :)
+  if (process.argv.includes('--postinstall) && fs.existsSync(outputDir)) return;
 
-transformer.generate(outputDir)
-  .catch(err => {
-    console.error(err);
-    process.exit(1);
-  });
+  // Clean up any old stuff lying around.
+  rimraf.sync(outputDir);
+  fs.mkdirSync(outputDir);
+
+  transformer.generate(outputDir)
+    .catch(err => {
+      console.error(err);
+      process.exit(1);
+    });
+}());
